(hds-2383) Move core story components to a folder
