+++ packages/react/src/components/cookieConsentCore/storyComponents/Banner.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 
-import { CookieConsentCore } from './cookieConsentCore';
-import { CreateProps } from './types';
+import { CookieConsentCore } from '../cookieConsentCore';
+import { CreateProps } from '../types';
 
 export const Banner = (props: CreateProps) => {
   useEffect(() => {
